pyproject.toml
[tool.poetry]
name = "webcolor"
version = "0.1.0"
description = "Official implementation of Generative Colorization of Structured Mobile Web Pages, WACV'23"
authors = ["Kotaro Kikuchi <kikuchi_kotaro_xa@cyberagent.co.jp>"]
license = "Apache-2.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = {url = "https://download.pytorch.org/whl/cu116/torch-1.12.1%2Bcu116-cp310-cp310-linux_x86_64.whl"}
dgl-cu116 = {url = "https://data.dgl.ai/wheels/dgl_cu116-0.9.1-cp310-cp310-manylinux1_x86_64.whl"}
h5py = "^3.7.0"
pytorch-lightning = {version = "^1.8.6", extras = ["extra"]}
tensorboard = "^2.11.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pysen = {version = "^0.10.2"}
pysen-ls = "^0.1.2"
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.982"
ipython = "^8.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py38"
isort_known_first_party = ["webcolor"]

[[tool.pysen.lint.mypy_targets]]
paths = ["src"]

[tool.pysen.lint.source]
excludes = [".venv"]
